+++ packages/cli-tools/bin/streamr-internal-operator-unstake.ts
@@ -5,9 +5,9 @@ import { StreamrClient, _operatorContractUtils } from '@streamr/sdk'
 import { createClientCommand } from '../src/command'
 
 createClientCommand(async (client: StreamrClient, operatorContractAddress: string, sponsorshipAddress: string) => {
-    const operatorContract = _operatorContractUtils.getOperatorContract(operatorContractAddress).connect(await client.getSigner())
     await _operatorContractUtils.unstake(
-        operatorContract,
+        await client.getSigner(),
+        operatorContractAddress,
         sponsorshipAddress
     )
 })