+++ packages/cli-tools/bin/streamr-internal-sponsorship-create.ts
@@ -12,17 +12,12 @@ interface Options extends BaseOptions {
 }
 
 createClientCommand(async (client: StreamrClient, streamId: string, options: Options) => {
-    if (client.getConfig().environment !== 'dev2') {
-        // currently the deploySponsorshipContract uses TEST_CHAIN_CONFIG and therefore only "dev2" is supported
-        // TODO add e.g. "environment" parameter to that function so that other environments are also supported 
-        console.error('only "dev2" environment is supported')
-        process.exit(1)
-    }
     const contract = await _operatorContractUtils.deploySponsorshipContract({
         streamId: toStreamID(streamId, toEthereumAddress(await client.getUserId())),
         earningsPerSecond: parseEther(options.earningsPerSecond),
         deployer: await client.getSigner(),
-        minOperatorCount: options.minOperatorCount
+        minOperatorCount: options.minOperatorCount,
+        environmentId: client.getConfig().environment
     })
     console.info(JSON.stringify({ address: await contract.getAddress() }, undefined, 4))
 })