+++ packages/sdk/src/Message.ts
@@ -1,5 +1,21 @@
 import { HexString, StreamID } from '@streamr/utils'
-import { SignatureType, StreamMessage } from './protocol/StreamMessage'
+import { StreamMessage } from './protocol/StreamMessage'
+import { SignatureType } from '@streamr/trackerless-network'
+import { IdentityMapping, KeyType } from './identity/IdentityMapping'
+
+// Lookup structure for converting SignatureType to KeyType string
+export type MessageSignatureType = KeyType | 'ECDSA_SECP256K1_LEGACY' | 'ERC_1271'
+const stringVersionsOfSignatureTypes: Record<number, MessageSignatureType> = {
+    // Read key pair SignatureTypes from IdentityMapping
+    ...Object.fromEntries(
+        Object.entries(IdentityMapping).map(
+            ([keyType, config]) => [config.signatureType, keyType as KeyType]
+        )
+    ),
+    // These special ones need to be added manually
+    [SignatureType.ECDSA_SECP256K1_LEGACY]: 'ECDSA_SECP256K1_LEGACY',
+    [SignatureType.ERC_1271]: 'ERC_1271',
+}
 
 /**
  * Represents a message in the Streamr Network.
@@ -40,7 +56,7 @@ export interface Message {
     /**
      * Signature method used to sign message.
      */
-    signatureType: 'LEGACY_SECP256K1' | 'SECP256K1' | 'ERC_1271'
+    signatureType: MessageSignatureType
 
     /**
      * Publisher of message.
@@ -63,17 +79,12 @@ export interface Message {
 
 export type MessageMetadata = Omit<Message, 'content'>
 
-function signatureTypeToString(signatureType: SignatureType): 'LEGACY_SECP256K1' | 'SECP256K1' | 'ERC_1271' {
-    switch (signatureType) {
-        case SignatureType.LEGACY_SECP256K1:
-            return 'LEGACY_SECP256K1'
-        case SignatureType.SECP256K1:
-            return 'SECP256K1'
-        case SignatureType.ERC_1271:
-            return 'ERC_1271'
-        default:
-            throw new Error(`Unknown signature type: ${signatureType}`)
+function signatureTypeToString(signatureType: SignatureType): MessageSignatureType {
+    const result = stringVersionsOfSignatureTypes[signatureType]
+    if (!result) {
+        throw new Error(`Unknown signature type: ${signatureType}`)
     }
+    return result
 }
 
 export const convertStreamMessageToMessage = (msg: StreamMessage): Message => {