+++ packages/trackerless-network/test/end-to-end/proxy-connections.test.ts
@@ -38,7 +38,7 @@ const MESSAGE: StreamMessage = {
             encryptionType: EncryptionType.NONE,
         }
     },
-    signatureType: SignatureType.SECP256K1,
+    signatureType: SignatureType.ECDSA_SECP256K1_EVM,
     signature: hexToBinary('0x1234')
 }