+++ packages/trackerless-network/test/unit/PlumTreeRpcLocal.test.ts
@@ -1,0 +1,62 @@
+import { ListeningRpcCommunicator, PeerDescriptor, toNodeId } from '@streamr/dht'
+import { PlumTreeRpcLocal } from '../../src/logic/plumtree/PlumTreeRpcLocal'
+import { createMockPeerDescriptor } from '../utils/utils'
+import { NodeList } from '../../src/logic/NodeList'
+import { ContentDeliveryRpcRemote } from '../../src/logic/ContentDeliveryRpcRemote'
+import { MockTransport } from '../utils/mock/MockTransport'
+import { ContentDeliveryRpcClient } from '../../generated/packages/trackerless-network/protos/NetworkRpc.client'
+import { PausedNeighbors } from '../../src/logic/plumtree/PausedNeighbors'
+
+describe('PlumTreeRpcLocal', () => {
+
+    let rpcLocal: PlumTreeRpcLocal
+    let localPeerDescriptor: PeerDescriptor
+    let rpcCommunicator: ListeningRpcCommunicator
+    let pausedNodes: PausedNeighbors
+    let neighbors: NodeList
+    let onMetadata: jest.Mock
+    let sendBuffer: jest.Mock
+
+    beforeEach(() => {
+        localPeerDescriptor = createMockPeerDescriptor()
+        rpcCommunicator = new ListeningRpcCommunicator('plumtree', new MockTransport())
+        neighbors = new NodeList(toNodeId(localPeerDescriptor), 4)
+        pausedNodes = new PausedNeighbors()
+        onMetadata = jest.fn()
+        sendBuffer = jest.fn()
+        rpcLocal = new PlumTreeRpcLocal(
+            neighbors,
+            pausedNodes,
+            onMetadata,
+            sendBuffer
+        )
+    })
+
+    it('pause neighbor', async () => {
+        const neighbor = createMockPeerDescriptor()
+        neighbors.add(new ContentDeliveryRpcRemote(localPeerDescriptor, neighbor, rpcCommunicator, ContentDeliveryRpcClient))
+        await rpcLocal.pauseNeighbor({ messageChainId: 'test' }, { incomingSourceDescriptor: neighbor } as any)
+        expect(pausedNodes.isPaused(toNodeId(neighbor), 'test')).toBe(true)
+    })
+
+    it('pause neighbor neighor not in neighbors', async () => {
+        const neighbor = createMockPeerDescriptor()
+        await rpcLocal.pauseNeighbor({ messageChainId: 'test' }, { incomingSourceDescriptor: neighbor } as any)
+        expect(pausedNodes.isPaused(toNodeId(neighbor), 'test')).toBe(false)
+    })
+
+    it('resume neighbor', async () => {
+        const neighbor = createMockPeerDescriptor()
+        neighbors.add(new ContentDeliveryRpcRemote(localPeerDescriptor, neighbor, rpcCommunicator, ContentDeliveryRpcClient))
+        await rpcLocal.pauseNeighbor({ messageChainId: 'test' }, { incomingSourceDescriptor: neighbor } as any)
+        expect(pausedNodes.isPaused(toNodeId(neighbor), 'test')).toBe(true)
+        rpcLocal.resumeNeighbor({ fromTimestamp: 0, messageChainId: 'test' }, { incomingSourceDescriptor: neighbor } as any)
+        expect(pausedNodes.isPaused(toNodeId(neighbor), 'test')).toBe(false)
+    })
+
+    it('send metadata', async () => {
+        const neighbor = createMockPeerDescriptor()
+        await rpcLocal.sendMetadata({ messageChainId: 'test' } as any, { incomingSourceDescriptor: neighbor } as any)
+        expect(onMetadata).toHaveBeenCalled()
+    })
+})